% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableVars.R, R/loadLTERdata.R
\name{availableVars.1D}
\alias{availableVars.1D}
\title{List of 1-D available variables from NTL datasets}
\usage{
availableVars.1D()

availableVars.1D()
}
\description{
List of 1-D available variables from NTL datasets

List of 1-D available variables from NTL datasets
}
