% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massCalc.R
\name{calcMass}
\alias{calcMass}
\title{Calculate weekly or annual mass in the lake}
\usage{
calcMass(df.interpolated, lakeAbr, time.res = "weekly", conversion = 1)
}
\arguments{
\item{df.interpolated}{dataframe of weekly interpolated data. Output of weeklyInterpolate()}

\item{lakeAbr}{Lake identification, string}

\item{time.res}{Time resolution. Options are 'weekly' (gives a daily mass at a weekly timestep) or 
'annual' (gives an annual mass averaged from the weekly data)}

\item{conversion}{Load unit conversion. Ex. if unit of var is µg/L and the mass will be output in µg.
A conversion factor of 1e6 will provide kg.}
}
\description{
Input weekly 1-m depth profiles to calculate depth integrated mass at weekly or annual timescales using lake bathymetry
}
