% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeTS.R
\name{decomposeTS}
\alias{decomposeTS}
\title{Decompose timeseries from weekly load data}
\usage{
decomposeTS(df.load)
}
\arguments{
\item{df}{dataframe of weekly load estimates}

\item{lakeAbr}{Lake identification, string}
}
\description{
Input load data
}
