% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLTERdata.R
\name{loadLTERchlorophyll.north}
\alias{loadLTERchlorophyll.north}
\title{Load LTER chlorophyll data from web}
\usage{
loadLTERchlorophyll.north()
}
\description{
Chlorophyll and phaeopigments are measured at our permanent sampling station
in the deepest part of each lake. Chlorophyll samples are collected from the
seven primary study lakes (Allequash, Big Muskellunge, Crystal, Sparkling, and
Trout lakes and bog lakes 27-02 [Crystal Bog], and 12-15 [Trout Bog]) in the
Trout Lake area at two to 10 depths depending on the lake and analyzed
spectrophotometrically. Sampling Frequency: fortnightly during ice-free season
- every 6 weeks during ice-covered season Number of sites: 7
}
