% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLTERdata.R
\name{loadLTERchlorophyll}
\alias{loadLTERchlorophyll}
\title{Load LTER chlorophyll data from web}
\usage{
loadLTERchlorophyll()
}
\description{
Downloads Chlorophyll of Northern Study Lakes 1981 - current
}
