% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLTERdata.R
\name{loadLTERchlorophyll.south}
\alias{loadLTERchlorophyll.south}
\title{Load LTER chlorophyll data from web}
\usage{
loadLTERchlorophyll.south()
}
\description{
Downloads Chlorophyll of Northern Study Lakes 1981 - current Chlorophyll is
measured at our permanent sampling station in the deepest part of the lake.
Chlorophyll samples are collected from the four primary study lakes in the
Madison area (Lakes Mendota, Monona, and Wingra and Fish Lake) at integrated
depths and discrete depths for spectrophotometric analysis and fluorometric
analysis. Due to a change in instruments starting in 2002 and lasting through
2007, chlorophyll analyses for the southern lakes had an uncorrectable bias,
and are not included in this dataset.  Analyses since then (2008 onward) have
been determined to not have this bias. Sampling Frequency: bi-weekly during
ice-free season from late March or early April through early September, then
every 4 weeks through late November; sampling is conducted usually once
during the winter (depending on ice conditions). Number of sites: 4
}
