% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLTERdata.R
\name{loadLTERnutrients}
\alias{loadLTERnutrients}
\title{Load LTER nutrient data from web}
\usage{
loadLTERnutrients()
}
\description{
Downloads Chemical Limnology of Primary Study Lakes: Nutrients, pH and Carbon 1981 - current
}
