% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLTERdata.R
\name{loadLTERphytoplankton}
\alias{loadLTERphytoplankton}
\title{Load LTER phytoplankton from web}
\usage{
loadLTERphytoplankton()
}
\description{
Phytoplankton samples for the 4 southern Wisconsin LTER lakes (Mendota,
Monona, Wingra, Fish) have been collected for analysis by LTER since 1995
(1996 Wingra, Fish) when the southern Wisconsin lakes were added to the North
Temperate Lakes LTER project. Samples are collected as a composite
whole-water sample and are preserved in gluteraldehyde. Composite sample
depths are 0-8 meters for Lake Mendota (to conform to samples collected and
analyzed since 1990 for a UW/DNR food web research study), and 0-2 meters for
the other three lakes. A tube sampler is used for the 0-8 m Lake Mendota
samples; samples for the other lakes are obtained by collecting water at
1-meter intervals using a Kemmerer water sampler and compositing the samples
in a bucket. Samples are taken in the deep hole region of each lake at the
same time and location as other limnological sampling. Phytoplankton samples
are analyzed by PhycoTech, Inc., a private lab specializing in phytoplankton
analyses (see data protocol for procedures). Samples for Wingra and Fish
lakes are archived but not routinely counted. Permanent slide mounts (3 per
sample) are prepared for all analyzed Mendota and Monona samples as well as 6
samples per year for Wingra and Fish; the slide mounts are archived at the
University of Wisconsin - Madison Zoology Museum. Phytoplankton are
identified to species using an inverted microscope (Utermohl technique) and
are reported as natural unit (i.e., colonies, filaments, or single cells)
densities per mL, cell densities per mL, and algal biovolume densities per
mL. Multiple entries for the same species on the same date may be due to
different variants or vegetative states - (e.g., colonial or attached vs.
free cell.) Biovolumes for individual cells of each species are determined
during the counting procedure by obtaining cell measurements needed to
calculate volumes for geometric solids (e.g., cylinders, spheres, truncated
cones) corresponding to actual cell shapes. Biovolume concentrations are then
computed by mulitplying the average cell biovolume by the cell densities in
the water sample. Note that one million cubicMicrometers of biovolume
PerMilliliter of water are equal to a biovolume concentration of one
cubicMillimeterPerMilliliter. Assuming a cell density equal to water, a
cubicMillimeterPerMilliliter of biovolume converts to a biomass concentration
of one milligramPerLiter. Sampling Frequency: bi-weekly during ice-free
season from late March or early April through early September, then every 4
weeks through late November; sampling is conducted usually once during the
winter (depending on ice conditions). Number of sites: 4 Several taxonomic
updates have been made to this dataset February 2013, see methods for
details.
}
