% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLTERdata.R
\name{loadLTERsecchi}
\alias{loadLTERsecchi}
\title{Load LTER secchi data from web}
\usage{
loadLTERsecchi()
}
\description{
Secchi disk depth is measured in the deepest part of each lake for the eleven primary lakes
(Allequash, Big Muskellunge, Crystal, Sparkling, Trout lakes, unnamed lakes 27-02 [Crystal Bog]
and 12-15 [Trout Bog], Fish, Mendota, Monona and Wingra). The disk is circular, 20 cm in diameter,
and has alternating black and white quadrants. It is lowered using a calibrated Kevlar rope to
minimize stretching. Readings are made on the shaded side of the boat both with and without the
aid of a plexiglass viewer. The points at which the disk disappears while being lowered and
reappears while being raised are averaged to determine Secchi depth. Auxiliary data include time
of day, air temperature, cloud cover, wave height, wind speed and direction and whether the lake
was ice covered on the sampledate. Sampling Frequency: fortnightly during ice-free season - every
6 weeks during ice-covered season for the northern lakes. The southern lakes are similar except
that sampling occurs monthly during the fall and typically only once during the winter (depending
on ice conditions). Number of sites: 11
}
