% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLTERdata.R
\name{loadLTERsnow}
\alias{loadLTERsnow}
\title{Load LTER snow data from web}
\usage{
loadLTERsnow()
}
\description{
Snow and ice depth are measured during the winter months on the eleven primary lakes
(Allequash, Big Muskellunge, Crystal, Sparkling, Trout lakes, unnamed lakes 27-02
[Crystal Bog] and 12-15 [Trout Bog], Fish, Mendota, Monona and Wingra). Sampling Frequency:
every 6 weeks during ice-covered season in the north and typically once during the winter
in the south Number of sites: 11
}
