% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLTERdata.R
\name{loadLTERzooplankton}
\alias{loadLTERzooplankton}
\title{Load LTER zooplankton from web}
\usage{
loadLTERzooplankton()
}
\description{
Zooplankton samples for the 4 southern Wisconsin LTER lakes (Mendota, Monona,
Wingra, Fish) have been collected for analysis by LTER since 1995 (1996
Wingra, Fish) when the southern Wisconsin lakes were added to the North
Temperate Lakes LTER project. Samples are collected as a vertical tow using an
80-micron mesh conical net with a 30-cm diameter opening (net mouth: net
length ratio = 1:3) consistent with sampling conducted by the Wisconsin Dept.
Natural Resources in prior years. Zooplankton tows are taken in the deep hole
region of each lake at the same time and location as other limnological
sampling; zooplankton samples are preserved in 70% ethanol for later
processing. Samples are usually collected with standard tow depths on most
dates (e.g., 20 meters for Lake Mendota) but not always, so tow depth is
recorded as a variate in the database. Crustacean species are identified and
counted for Mendota and Monona and body lengths are recorded for a portion of
each species identified (see data protocol for counting procedure); samples
for Wingra and Fish lakes are archived but not routinely counted. Numerical
densities for Mendota and Monona zooplankton samples are reported in the
database as number or organisms per square meter without correcting for net
efficiency. [Net efficiency varies from a maximum of about 70% under clear
water conditions; net efficiency declines when algal blooms are dense
(Lathrop, R.C. 1998. Water clarity responses to phosphorus and Daphnia in Lake
Mendota. Ph.D. Thesis, University of Wisconsin-Madison.)] Organism densities
in number per cubic meter can be obtained by dividing the reported
square-meter density by the tow depth, although adjustments for the oxygenated
depth zone during the summer and early fall stratified season is required to
obtain realistic zooplankton volumetric densities in the lake's surface
waters. Biomass densities can be calculated using literature formulas for
converting organism body lengths reported in the database to body masses.
Sampling Frequency: bi-weekly during ice-free season from late March or early
April through early September, then every 4 weeks through late November;
sampling is conducted usually once during the winter (depending on ice
conditions). Number of sites: 4 Note: for a period between approximately 2011
and 2015, a calculation error caused density values to be significantly
greater than they should have been for the entire dataset. That issue has been
corrected.
}
