% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateFunctions.R
\name{weeklyInterpolate.1D}
\alias{weeklyInterpolate.1D}
\title{Interpolate LTER data to weekly}
\usage{
weeklyInterpolate.1D(lakeAbr, var)
}
\arguments{
\item{lakeAbr}{Lake identification, string}

\item{var}{Variable of interest. Use availableVars() to see available variables.}
}
\description{
Interpolates dataframe of 1-D observations (secchi, light extinction) to weekly timestep.
}
