% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateFunctions.R
\name{weeklyInterpolate}
\alias{weeklyInterpolate}
\title{Interpolate LTER data to weekly}
\usage{
weeklyInterpolate(
  lakeAbr,
  var,
  dataset,
  maxdepth,
  constrainMethod = "zero",
  setThreshold = 0.1,
  printFigs = FALSE
)
}
\arguments{
\item{lakeAbr}{Lake identification, string}

\item{var}{Variable of interest. Use availableVars() to see available variables.}

\item{dataset}{NTL dataset with the variable of interest downloaded via loadLTER...() functions}

\item{maxdepth}{Maximum depth of lake}

\item{constrainMethod}{Options to constrain interpolation.
Options are 'zero' (cannot go below zero, default) and 'range' (can not go beyond range of observed data)}

\item{setThreshold}{Threshold of RMSE for interpolation. Use units of variable. If the function returns many 
predict.lm warnings, the threshold should be increased.}

\item{printFigs}{Output individual profiles of interpolation? Options TRUE or FALSE (default)}
}
\value{
returns a dataframe with three columns, data, depth, and var
#' @examples
#Interpolate TP data
NTLnutrients = loadLTERnutrients()
weeklyInterpolate.1D(lakeAbr = 'ME',var = 'drp_sloh',dataset = NTLnutrients)
}
\description{
Interpolates dataframe of observations to weekly timestep at 1 m depth intervals.
Uses the premise that nutrient profiles should match temperature profiles
}
