% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateTempFunctions.R
\name{weeklyTempInterpolate}
\alias{weeklyTempInterpolate}
\title{Interpolate temperature data to weekly}
\usage{
weeklyTempInterpolate(
  lakeAbr,
  maxdepth,
  dataset = LTERtemp,
  constrainMethod = "zero"
)
}
\arguments{
\item{lakeAbr}{Lake identification, string}

\item{maxdepth}{Maximum depth of lake}

\item{dataset}{temperature dataset. Default woudl be LTERtemp loaded via loadLTERtemp()}

\item{constrainMethod}{Options to constrain interpolation.
Options are 'zero' (cannot go below zero, cannot go 1.5x above max, default) 
and 'range' (can not go beyond range of observed data)}
}
\value{
returns a list 
[[1]]: observation data
[[2]]: dataframe with three columns, data, depth, and var
#' @examples
#Interpolate temp data
LTERtemp = loadLTERtemp()
weeklyTempInterpolate(lakeAbr = 'ME', maxdepth = 24, dataset = LTERtemp)
}
\description{
Interpolates dataframe of observations to weekly timestep at 1 m depth intervals.
Uses simple linear interpolation
}
