% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTER_loadCalc.R
\name{calcLoad}
\alias{calcLoad}
\title{Calculate weekly or annual load}
\usage{
calcLoad(df, lakeAbr, time.res = "weekly")
}
\arguments{
\item{df}{dataframe of weekly interpolated data. Output of weeklyInterpolate()}

\item{lakeAbr}{Lake identification, string}

\item{time.res}{Time resolution. Options are 'weekly' (default) or 'annual'}
}
\description{
Input weekly 1-m depth profiles to calculate depth integrated load at weekly or annual timescales
}
