% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeTS.R
\name{decomposeTS}
\alias{decomposeTS}
\title{Decompose timeseries from weekly mass data}
\usage{
decomposeTS(df.mass, lakeAbr, var)
}
\arguments{
\item{df.mass}{dataframe of weekly mass estimates}

\item{lakeAbr}{Lake identification, string}

\item{var}{Variable of interest. Use availableVars() to see available variables.}
}
\description{
Input mass data
}
