% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeseries.R
\name{plotTimeseries}
\alias{plotTimeseries}
\title{Plot interpolate heatmap of entire timeseries}
\usage{
plotTimeseries(df, saveFig = FALSE)
}
\arguments{
\item{df}{dataframe of weekly interpolated data. Output of weeklyInterpolate()}
}
\description{
Plot interpolate heatmap of entire timeseries
}
