% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeseries.R
\name{plotTimeseries.year}
\alias{plotTimeseries.year}
\title{Plot interpolate heatmap of single year}
\usage{
plotTimeseries.year(
  df.interpolated,
  observations,
  var,
  chooseYear,
  saveFig = FALSE
)
}
\arguments{
\item{df.interpolated}{dataframe of weekly interpolated data. Output of weeklyInterpolate()}

\item{observations}{dataframe of observed data. Output of weeklyInterpolate()}

\item{var}{Variable of interest. Use availableVars() to see available variables.}

\item{chooseYear}{year of interest}
}
\description{
Plot interpolate heatmap of single year
}
