% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeseries.R
\name{plotTimeseries.year}
\alias{plotTimeseries.year}
\title{Plot interpolate heatmap of single year}
\usage{
plotTimeseries.year(df, observations, chooseYear, saveFig = FALSE)
}
\arguments{
\item{df}{dataframe of weekly interpolated data. Output of weeklyInterpolate()}

\item{observations}{dataframe of observed data. Output of weeklyInterpolate()}

\item{chooseYear}{year of interest}
}
\description{
Plot interpolate heatmap of single year
}
