% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeseries.R
\name{plotTimeseries.year}
\alias{plotTimeseries.year}
\title{Plot interpolate heatmap of single year}
\usage{
plotTimeseries.year(
  df.interpolated,
  observations = NULL,
  var,
  chooseYear,
  binsize = NULL,
  legend.title = NULL,
  saveFig = FALSE
)
}
\arguments{
\item{df.interpolated}{dataframe of weekly interpolated data. Output of weeklyInterpolate()}

\item{observations}{To plot observed data, provide a dataframe of observed data. Output of weeklyInterpolate().
If NULL (default) no points plotted}

\item{var}{Variable of interest. Use availableVars() to see available variables.}

\item{chooseYear}{year of interest}

\item{binsize}{Size of legend bins. Defaults to 1/20th of maximum value.}

\item{legend.title}{Choice to specify legend title}

\item{saveFig}{If TRUE, saves a png to the working directory (default = FALSE)}
}
\description{
Plot interpolate heatmap of single year
}
