% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateFunctions.R
\name{weeklyInterpolate.1D}
\alias{weeklyInterpolate.1D}
\title{Interpolate LTER data to weekly}
\usage{
weeklyInterpolate.1D(lakeAbr, var, dataset)
}
\arguments{
\item{lakeAbr}{Lake identification, string}

\item{var}{Variable of interest. Use availableVars() to see available variables.}

\item{dataset}{NTL dataset with the variable of interest downloaded via loadLTER...() functions}
}
\value{
returns a dataframe with three columns, data, depth, and var
}
\description{
Interpolates dataframe of 1-D observations (secchi, light extinction) to weekly timestep.
}
\examples{
#Interpolate Secchi data
NTLsecchi = loadLTERsecchi()
weeklyInterpolate.1D(lakeAbr = 'ME',var = 'secnview',dataset = NTLsecchi)
}
